fix: refactor module structure
current implementation of pagination in only applicable to alchemy backend so it should be part of the backend
