+++ venom_resource/backends/alchemy/__init__.py
@@ -1,0 +1,1 @@
+from .resource import SQLAlchemyResource
